refactor: split large server and activity files into focused modules
Move mcp schema/validate/shell helpers, HTTP routing, tool-call activity pagination, and ActivityRunDialog UI chrome out of god files so navigation stays local without behavior changes.
