+++ src/db/tool-calls.ts
@@ -4,6 +4,7 @@ import {
 	isActivityRunOpener,
 } from "../shared/activity-run-boundary";
 import type { ToolCallRecord, ToolCallStats } from "../types/database";
+import { listRecentToolCalls } from "./tool-calls-activity-page";
 
 export const TOOL_CALLS_PER_PROJECT_CAP = 10_000;
 export const TOOL_CALLS_PAGE_SIZE_DEFAULT = 50;
@@ -77,34 +78,6 @@ export type ActivityCorrelationLookup = {
 };
 
 export { isActivityRunCloser, isActivityRunOpener };
-
-/** Cap how far we walk older rows to complete a cut-off run. */
-const RUN_BOUNDARY_EXPAND_MAX = TOOL_CALLS_PER_PROJECT_CAP;
-
-type ActivityPageUnit = {
-	unitKey: string;
-	kind: "conversation" | "orphan";
-	maxStarted: number;
-	traceCount: number;
-};
-
-function compareActivityUnitsDesc(
-	a: ActivityPageUnit,
-	b: ActivityPageUnit,
-): number {
-	if (b.maxStarted !== a.maxStarted) return b.maxStarted - a.maxStarted;
-	return b.unitKey.localeCompare(a.unitKey);
-}
-
-function isActivityUnitBeforeCursor(
-	unit: ActivityPageUnit,
-	cursor: { maxStarted: number; unitKey: string },
-): boolean {
-	if (unit.maxStarted !== cursor.maxStarted) {
-		return unit.maxStarted < cursor.maxStarted;
-	}
-	return unit.unitKey.localeCompare(cursor.unitKey) < 0;
-}
 
 export class ToolCallsRepo {
 	constructor(private db: Database) {}
@@ -232,515 +205,11 @@ export class ToolCallsRepo {
 		return row.n;
 	}
 
-	/** Rows strictly older than `(startedAt, id)` in newest-first order. */
-	private listBefore(
-		projectId: string,
-		startedAt: number,
-		id: string,
-		limit: number,
-	): ToolCallRecord[] {
-		return this.db
-			.query(
-				`SELECT * FROM tool_calls
-         WHERE project_id = ?
-           AND (started_at < ? OR (started_at = ? AND id < ?))
-         ORDER BY started_at DESC, id DESC
-         LIMIT ?`,
-			)
-			.all(projectId, startedAt, startedAt, id, limit) as ToolCallRecord[];
-	}
-
-	private hasRowBefore(
-		projectId: string,
-		startedAt: number,
-		id: string,
-	): boolean {
-		const row = this.db
-			.query(
-				`SELECT 1 AS ok FROM tool_calls
-         WHERE project_id = ?
-           AND (started_at < ? OR (started_at = ? AND id < ?))
-         LIMIT 1`,
-			)
-			.get(projectId, startedAt, startedAt, id) as { ok: number } | null;
-		return row != null;
-	}
-
-	/**
-	 * If the page's oldest row is mid-generation (or mid-heuristic-run), pull
-	 * older rows until the generation / run opener. Prefer provider generation
-	 * ids when present; fall back to prompt/stop heuristics.
-	 */
-	private expandOlderToRunBoundary(
-		projectId: string,
-		page: ToolCallRecord[],
-	): ToolCallRecord[] {
-		if (page.length === 0) return page;
-		const oldest = page[page.length - 1]!;
-
-		if (oldest.generation_id) {
-			return this.expandOlderMatching(
-				projectId,
-				page,
-				(row) => row.generation_id === oldest.generation_id,
-			);
-		}
-
-		if (isActivityRunOpener(oldest)) return page;
-
-		const expanded = [...page];
-		let walked = 0;
-		let foundBoundary = false;
-
-		while (walked < RUN_BOUNDARY_EXPAND_MAX) {
-			const tip = expanded[expanded.length - 1]!;
-			const batchSize = Math.min(50, RUN_BOUNDARY_EXPAND_MAX - walked);
-			const batch = this.listBefore(
-				projectId,
-				tip.started_at,
-				tip.id,
-				batchSize,
-			);
-			if (batch.length === 0) {
-				break;
-			}
-
-			for (const row of batch) {
-				walked += 1;
-				if (isActivityRunCloser(row)) {
-					foundBoundary = true;
-					break;
-				}
-				expanded.push(row);
-				if (isActivityRunOpener(row)) {
-					foundBoundary = true;
-					break;
-				}
-				if (walked >= RUN_BOUNDARY_EXPAND_MAX) break;
-			}
-
-			if (foundBoundary) break;
-			if (batch.length < batchSize) break;
-			if (walked >= RUN_BOUNDARY_EXPAND_MAX) break;
-		}
-
-		return foundBoundary ? expanded : page;
-	}
-
-	/** Pull older rows while `matches` stays true (e.g. same generation_id). */
-	private expandOlderMatching(
-		projectId: string,
-		page: ToolCallRecord[],
-		matches: (row: ToolCallRecord) => boolean,
-	): ToolCallRecord[] {
-		const expanded = [...page];
-		let walked = 0;
-
-		while (walked < RUN_BOUNDARY_EXPAND_MAX) {
-			const tip = expanded[expanded.length - 1]!;
-			const batchSize = Math.min(50, RUN_BOUNDARY_EXPAND_MAX - walked);
-			const batch = this.listBefore(
-				projectId,
-				tip.started_at,
-				tip.id,
-				batchSize,
-			);
-			if (batch.length === 0) break;
-
-			let hitMismatch = false;
-			for (const row of batch) {
-				walked += 1;
-				if (!matches(row)) {
-					hitMismatch = true;
-					break;
-				}
-				expanded.push(row);
-				if (walked >= RUN_BOUNDARY_EXPAND_MAX) break;
-			}
-
-			if (hitMismatch) break;
-			if (batch.length < batchSize) break;
-			if (walked >= RUN_BOUNDARY_EXPAND_MAX) break;
-		}
-
-		return expanded;
-	}
-
 	listRecent(
 		projectId: string,
 		options: ToolCallListOptions = {},
 	): ToolCallListResult {
-		const traceBudget = Math.max(
-			1,
-			Math.min(
-				options.limit ?? TOOL_CALLS_PAGE_SIZE_DEFAULT,
-				TOOL_CALLS_PAGE_SIZE_MAX,
-			),
-		);
-		const beforeStartedAt = options.beforeStartedAt ?? options.before ?? null;
-		const beforeId = options.beforeId ?? null;
-		const sessionId = options.sessionId?.trim() || null;
-		const conversationId = options.conversationId?.trim() || null;
-		const generationId = options.generationId?.trim() || null;
-
-		if (generationId) {
-			const fetched = this.db
-				.query(
-					`SELECT * FROM tool_calls
-           WHERE project_id = ? AND generation_id = ?
-           ORDER BY started_at ASC, id ASC`,
-				)
-				.all(projectId, generationId) as ToolCallRecord[];
-
-			return {
-				calls: fetched,
-				total: fetched.length,
-				hasMore: false,
-			};
-		}
-
-		if (sessionId) {
-			let fetched: ToolCallRecord[];
-			if (beforeStartedAt != null && beforeId) {
-				fetched = this.db
-					.query(
-						`SELECT * FROM tool_calls
-           WHERE project_id = ? AND session_id = ?
-             AND (started_at < ? OR (started_at = ? AND id < ?))
-           ORDER BY started_at DESC, id DESC
-           LIMIT ?`,
-					)
-					.all(
-						projectId,
-						sessionId,
-						beforeStartedAt,
-						beforeStartedAt,
-						beforeId,
-						traceBudget + 1,
-					) as ToolCallRecord[];
-			} else if (beforeStartedAt != null) {
-				fetched = this.db
-					.query(
-						`SELECT * FROM tool_calls
-           WHERE project_id = ? AND session_id = ?
-             AND started_at < ?
-           ORDER BY started_at DESC, id DESC
-           LIMIT ?`,
-					)
-					.all(
-						projectId,
-						sessionId,
-						beforeStartedAt,
-						traceBudget + 1,
-					) as ToolCallRecord[];
-			} else {
-				fetched = this.db
-					.query(
-						`SELECT * FROM tool_calls
-           WHERE project_id = ? AND session_id = ?
-           ORDER BY started_at DESC, id DESC
-           LIMIT ?`,
-					)
-					.all(projectId, sessionId, traceBudget + 1) as ToolCallRecord[];
-			}
-
-			const overflow = fetched.length > traceBudget;
-			if (overflow) fetched.pop();
-
-			return {
-				calls: fetched,
-				total: this.countForSession(projectId, sessionId),
-				hasMore: overflow,
-			};
-		}
-
-		if (conversationId) {
-			const fetched = this.listToolCallsForConversation(
-				projectId,
-				conversationId,
-			);
-
-			return {
-				calls: fetched,
-				total: fetched.length,
-				hasMore: false,
-			};
-		}
-
-		return this.listRecentByActivityUnits(
-			projectId,
-			traceBudget,
-			beforeStartedAt,
-			beforeId,
-		);
-	}
-
-	/**
-	 * All traces for a provider conversation, including Cursor rows stored under
-	 * an agent-session conversation_id that share a generation with the chat id.
-	 */
-	private listToolCallsForConversation(
-		projectId: string,
-		conversationId: string,
-	): ToolCallRecord[] {
-		const chatId = conversationId.trim();
-		const transcriptNeedle = `%agent-transcripts/${chatId}/%`;
-
-		const seedRows = this.db
-			.query(
-				`SELECT * FROM tool_calls
-         WHERE project_id = ?
-           AND (
-             conversation_id = ?
-             OR (attributes_json IS NOT NULL AND attributes_json LIKE ?)
-           )`,
-			)
-			.all(projectId, chatId, transcriptNeedle) as ToolCallRecord[];
-
-		const allowedConversationIds = new Set<string>([chatId]);
-		const generationIds = new Set<string>();
-		for (const row of seedRows) {
-			const gen = row.generation_id?.trim();
-			if (gen) generationIds.add(gen);
-		}
-
-		for (const genId of generationIds) {
-			const genRows = this.db
-				.query(
-					`SELECT * FROM tool_calls WHERE project_id = ? AND generation_id = ?`,
-				)
-				.all(projectId, genId) as ToolCallRecord[];
-
-			const promptConvIds = new Set<string>();
-			for (const row of genRows) {
-				if (row.kind === "prompt" && row.conversation_id?.trim()) {
-					promptConvIds.add(row.conversation_id.trim());
-				}
-			}
-			if (promptConvIds.size > 1) continue;
-
-			if (!genRows.some((row) => row.conversation_id?.trim() === chatId)) {
-				continue;
-			}
-
-			for (const row of genRows) {
-				const cid = row.conversation_id?.trim();
-				if (!cid || cid === chatId || allowedConversationIds.has(cid)) continue;
-
-				// Do not pull tool rows stored under another chat conversation id
-				// that owns prompts elsewhere, even when generation_id collides.
-				const ownsPromptElsewhere = this.db
-					.query(
-						`SELECT 1 FROM tool_calls
-             WHERE project_id = ? AND conversation_id = ? AND kind = 'prompt'
-             LIMIT 1`,
-					)
-					.get(projectId, cid);
-				if (ownsPromptElsewhere) continue;
-
-				allowedConversationIds.add(cid);
-			}
-		}
-
-		if (allowedConversationIds.size === 1 && seedRows.length === 0) {
-			return [];
-		}
-
-		const placeholders = [...allowedConversationIds]
-			.map(() => "?")
-			.join(", ");
-		return this.db
-			.query(
-				`SELECT * FROM tool_calls
-         WHERE project_id = ?
-           AND conversation_id IN (${placeholders})
-         ORDER BY started_at DESC, id DESC`,
-			)
-			.all(projectId, ...allowedConversationIds) as ToolCallRecord[];
-	}
-
-	/**
-	 * Paginate activity by full conversations (or orphan run units), packing
-	 * units until the trace budget is reached. A single unit larger than the
-	 * budget is returned alone.
-	 */
-	private listRecentByActivityUnits(
-		projectId: string,
-		traceBudget: number,
-		beforeStartedAt: number | null,
-		beforeId: string | null,
-	): ToolCallListResult {
-		const cursor =
-			beforeStartedAt != null && beforeId
-				? this.resolveActivityPageCursor(
-						projectId,
-						beforeStartedAt,
-						beforeId,
-					)
-				: beforeStartedAt != null
-					? { maxStarted: beforeStartedAt, unitKey: "" }
-					: null;
-
-		const units = this.listActivityPageUnits(projectId, cursor);
-		const selected = this.selectUnitsByTraceBudget(units, traceBudget);
-		const calls = this.fetchTracesForActivityUnits(projectId, selected);
-
-		return {
-			calls,
-			total: this.count(projectId),
-			hasMore: selected.length < units.length,
-		};
-	}
-
-	private listActivityPageUnits(
-		projectId: string,
-		cursor?: { maxStarted: number; unitKey: string } | null,
-	): ActivityPageUnit[] {
-		const conversationRows = this.db
-			.query(
-				`SELECT conversation_id AS unit_key, MAX(started_at) AS max_started, COUNT(*) AS trace_count
-         FROM tool_calls
-         WHERE project_id = ? AND conversation_id IS NOT NULL
-         GROUP BY conversation_id`,
-			)
-			.all(projectId) as Array<{
-			unit_key: string;
-			max_started: number;
-			trace_count: number;
-		}>;
-
-		const orphanRows = this.db
-			.query(
-				`SELECT COALESCE(generation_id, id) AS unit_key, MAX(started_at) AS max_started, COUNT(*) AS trace_count
-         FROM tool_calls
-         WHERE project_id = ? AND conversation_id IS NULL
-         GROUP BY COALESCE(generation_id, id)`,
-			)
-			.all(projectId) as Array<{
-			unit_key: string;
-			max_started: number;
-			trace_count: number;
-		}>;
-
-		let units: ActivityPageUnit[] = [
-			...conversationRows.map((row) => ({
-				unitKey: row.unit_key,
-				kind: "conversation" as const,
-				maxStarted: Number(row.max_started),
-				traceCount: Number(row.trace_count),
-			})),
-			...orphanRows.map((row) => ({
-				unitKey: row.unit_key,
-				kind: "orphan" as const,
-				maxStarted: Number(row.max_started),
-				traceCount: Number(row.trace_count),
-			})),
-		];
-
-		units.sort(compareActivityUnitsDesc);
-
-		if (cursor) {
-			units = units.filter((unit) => isActivityUnitBeforeCursor(unit, cursor));
-		}
-
-		return units;
-	}
-
-	private resolveActivityPageCursor(
-		projectId: string,
-		beforeStartedAt: number,
-		beforeId: string,
-	): { maxStarted: number; unitKey: string } {
-		const row = this.get(beforeId);
-		if (!row || row.project_id !== projectId) {
-			return { maxStarted: beforeStartedAt, unitKey: "" };
-		}
-
-		if (row.conversation_id) {
-			const summary = this.db
-				.query(
-					`SELECT MAX(started_at) AS max_started FROM tool_calls
-           WHERE project_id = ? AND conversation_id = ?`,
-				)
-				.get(projectId, row.conversation_id) as { max_started: number };
-			return {
-				maxStarted: Number(summary.max_started),
-				unitKey: row.conversation_id,
-			};
-		}
-
-		const orphanKey = row.generation_id ?? row.id;
-		const summary = this.db
-			.query(
-				`SELECT MAX(started_at) AS max_started FROM tool_calls
-         WHERE project_id = ? AND conversation_id IS NULL
-           AND (generation_id = ? OR (generation_id IS NULL AND id = ?))`,
-			)
-			.get(projectId, orphanKey, orphanKey) as { max_started: number } | null;
-
-		return {
-			maxStarted: summary
-				? Number(summary.max_started)
-				: beforeStartedAt,
-			unitKey: orphanKey,
-		};
-	}
-
-	private selectUnitsByTraceBudget(
-		units: ActivityPageUnit[],
-		traceBudget: number,
-	): ActivityPageUnit[] {
-		const selected: ActivityPageUnit[] = [];
-		let accumulated = 0;
-
-		for (const unit of units) {
-			if (selected.length === 0) {
-				selected.push(unit);
-				accumulated += unit.traceCount;
-				if (unit.traceCount > traceBudget) break;
-				continue;
-			}
-			if (accumulated + unit.traceCount > traceBudget) break;
-			selected.push(unit);
-			accumulated += unit.traceCount;
-		}
-
-		return selected;
-	}
-
-	private fetchTracesForActivityUnits(
-		projectId: string,
-		units: ActivityPageUnit[],
-	): ToolCallRecord[] {
-		const calls: ToolCallRecord[] = [];
-		for (const unit of units) {
-			if (unit.kind === "conversation") {
-				const rows = this.db
-					.query(
-						`SELECT * FROM tool_calls
-             WHERE project_id = ? AND conversation_id = ?
-             ORDER BY started_at DESC, id DESC`,
-					)
-					.all(projectId, unit.unitKey) as ToolCallRecord[];
-				calls.push(...rows);
-			} else {
-				const rows = this.db
-					.query(
-						`SELECT * FROM tool_calls
-             WHERE project_id = ? AND conversation_id IS NULL
-               AND (generation_id = ? OR (generation_id IS NULL AND id = ?))
-             ORDER BY started_at DESC, id DESC`,
-					)
-					.all(projectId, unit.unitKey, unit.unitKey) as ToolCallRecord[];
-				calls.push(...rows);
-			}
-		}
-
-		calls.sort((a, b) => {
-			if (b.started_at !== a.started_at) return b.started_at - a.started_at;
-			return b.id.localeCompare(a.id);
-		});
-		return calls;
+		return listRecentToolCalls(this.db, projectId, options);
 	}
 
 	/**
