+++ src/server/configure-routes.ts
@@ -6,13 +6,16 @@ import {
 import { logger } from "../shared/logger";
 import { detectCapabilitiesFile } from "../shared/paths";
 import { trustStdioServers } from "../shared/stdio-allowlist";
-import { resolveVariablesInObject } from "../shared/variable-resolver";
 import { projectUiUrl } from "../shared/ui-urls";
-import { extractAllVariables } from "../shared/variable-resolver";
+import {
+	extractAllVariables,
+	resolveVariablesInObject,
+} from "../shared/variable-resolver";
 import type { Capabilities } from "../types/capabilities";
 import type { OAuth2Config } from "../types/oauth";
 import type { CapabilitiesFileWatcher } from "./capabilities-watcher";
 import type { CapaMCPServer, ValidationProgressEvent } from "./mcp-handler";
+import type { McpServerStateManager } from "./mcp-server-state";
 import { OAuth2Manager } from "./oauth-manager";
 import {
 	type OAuth2ServerEntry,
@@ -24,7 +27,6 @@ import {
 	loadEffectiveCapabilities,
 } from "./resolve-effective-capabilities";
 import type { SessionManager } from "./session-manager";
-import type { McpServerStateManager } from "./mcp-server-state";
 
 const JSON_HEADERS = { "Content-Type": "application/json" };
 
@@ -474,4 +476,21 @@ export async function handleProjectConfigure(
 			"Cache-Control": "no-cache",
 		},
 	});
+}
+
+/**
+ * Dispatcher for `/api/projects/:id/configure`.
+ * Returns null if the path is not the configure route.
+ */
+export async function dispatchConfigure(
+	deps: ConfigureRouteDeps,
+	path: string,
+	method: string,
+	request: Request,
+): Promise<Response | null> {
+	const configMatch = path.match(/^\/api\/projects\/([^/]+)\/configure$/);
+	if (configMatch && method === "POST") {
+		return handleProjectConfigure(deps, configMatch[1], request);
+	}
+	return null;
 }
